MovieDB/requirements.txt
adium-theme-ubuntu==0.3.4
apt-xapian-index==0.45
chardet==2.0.1
colorama==0.2.5
command-not-found==0.3
debtagshw==0.1
defer==1.0.6
dirspec==13.10
Django==1.8.4
django-filter==0.11.0
django-tastypie==0.14.1
djangorestframework==3.2.4
dnspython==1.11.1
duplicity==0.6.23
gunicorn==19.3.0
gyp==0.1
html5lib==0.999
httplib2==0.8
jsonfield==1.0.3
lockfile==0.8
lxml==3.3.3
Markdown==2.6.11
oauthlib==0.6.1
oneconf==0.3.7
openshot==1.4.3
PAM==0.4.2
pbr==4.0.4
pexpect==3.1
Pillow==2.3.0
piston-mini-client==0.7.5
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pygobject==3.12.0
pyOpenSSL==0.13
pyserial==2.6
pysmbc==1.0.14.1
python-apt===0.9.3.5ubuntu1
python-dateutil==2.7.3
python-debian===0.1.21-nmu2ubuntu2
python-mimeparse==0.1.4
pyxdg==0.25
reportlab==3.0
requests==2.2.1
sessioninstaller==0.0.0
simplejson==3.8.0
six==1.5.2
software-center-aptd-plugins==0.0.0
ssh-import-id==3.21
stevedore==1.28.0
system-service==0.1.6
Twisted-Core==13.2.0
Twisted-Web==13.2.0
unity-lens-photos==1.0
urllib3==1.7.1
vboxapi==1.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
wheel==0.24.0
wxPython==2.8.12.1
wxPython-common==2.8.12.1
xdiagnose===3.6.3build2
zope.interface==4.0.5
